build: userData功能调整到frame模块
