+++ frame/src/main/java/com/tlcsdm/frame/SampleBase.java
@@ -26,22 +26,34 @@
  */
 package com.tlcsdm.frame;
 
+import java.io.File;
+import java.time.LocalDate;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Objects;
+import java.util.ServiceLoader;
+
+import com.tlcsdm.core.javafx.util.FxXmlUtil;
 import com.tlcsdm.frame.util.I18nUtils;
+
+import cn.hutool.core.util.StrUtil;
 import javafx.application.Application;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.CheckBox;
+import javafx.scene.control.DatePicker;
 import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.Separator;
 import javafx.scene.control.SplitPane;
+import javafx.scene.control.TextInputControl;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.TextFlow;
+import javafx.stage.DirectoryChooser;
+import javafx.stage.FileChooser;
 import javafx.stage.Stage;
-
-import java.util.Objects;
-import java.util.ServiceLoader;
 
 /**
  * A base class for samples - it is recommended that they extend this class
@@ -50,161 +62,228 @@ import java.util.ServiceLoader;
  */
 public abstract class SampleBase extends Application implements Sample {
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void start(Stage primaryStage) {
-        ServiceLoader<FXSamplerConfiguration> configurationServiceLoader = ServiceLoader
-                .load(FXSamplerConfiguration.class);
+	protected Map<String, Object> userData = new LinkedHashMap<>();
 
-        primaryStage.setTitle(getSampleName());
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public void start(Stage primaryStage) {
+		ServiceLoader<FXSamplerConfiguration> configurationServiceLoader = ServiceLoader
+				.load(FXSamplerConfiguration.class);
 
-        Scene scene = new Scene((Parent) buildSample(this, primaryStage), 800, 800);
-        scene.getStylesheets()
-                .add(Objects.requireNonNull(SampleBase.class.getResource("/fxsampler/fxsampler.css")).toExternalForm());
-        for (FXSamplerConfiguration fxsamplerConfiguration : configurationServiceLoader) {
-            String stylesheet = fxsamplerConfiguration.getSceneStylesheet();
-            if (stylesheet != null) {
-                scene.getStylesheets().add(stylesheet);
-            }
-        }
-        primaryStage.setScene(scene);
-        primaryStage.show();
-    }
+		primaryStage.setTitle(getSampleName());
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean isVisible() {
-        return true;
-    }
+		Scene scene = new Scene((Parent) buildSample(this, primaryStage), 800, 800);
+		scene.getStylesheets()
+				.add(Objects.requireNonNull(SampleBase.class.getResource("/fxsampler/fxsampler.css")).toExternalForm());
+		for (FXSamplerConfiguration fxsamplerConfiguration : configurationServiceLoader) {
+			String stylesheet = fxsamplerConfiguration.getSceneStylesheet();
+			if (stylesheet != null) {
+				scene.getStylesheets().add(stylesheet);
+			}
+		}
+		primaryStage.setScene(scene);
+		primaryStage.show();
+	}
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public Node getControlPanel() {
-        return null;
-    }
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public boolean isVisible() {
+		return true;
+	}
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public double getControlPanelDividerPosition() {
-        return 0.6;
-    }
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public Node getControlPanel() {
+		return null;
+	}
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getSampleDescription() {
-        return "";
-    }
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public double getControlPanelDividerPosition() {
+		return 0.6;
+	}
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getProjectName() {
-        return "ControlsFX";
-    }
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public String getSampleDescription() {
+		return "";
+	}
 
-    @Override
-    public void initialize() {
-        // do nothing
-    }
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public String getProjectName() {
+		return "ControlsFX";
+	}
 
-    /**
-     * Utility method to create the default look for samples.
-     */
-    public static Node buildSample(Sample sample, Stage stage) {
-        SplitPane splitPane = new SplitPane();
-        // we guarantee that the build order is panel then control panel.
-        final Node samplePanel = sample.getPanel(stage);
-        final Node controlPanel = sample.getControlPanel();
-        sample.initialize();
-        splitPane.setDividerPosition(0, sample.getControlPanelDividerPosition());
+	@Override
+	public void initialize() {
+		initializeUserData();
+	}
 
-        if (samplePanel != null) {
-            samplePanel.getStyleClass().add("center-panel");
-            splitPane.getItems().add(samplePanel);
-        }
+	/**
+	 * Because initialize() is called after getPanel() so userData needs to be
+	 * initialized in getPanel()
+	 */
+	protected void initializeUserData() {
+		if (!FxXmlUtil.hasKey(getSampleXmlPrefix(), "id")) {
+			return;
+		}
+		userData.forEach((key, value) -> {
+			String k = getSampleXmlPrefix() + "." + key;
+			String val = FxXmlUtil.get(k, "");
+			if (value instanceof FileChooser v) {
+				if (!StrUtil.isEmpty(val)) {
+					v.setInitialDirectory(new File(val));
+				}
+			} else if (value instanceof DirectoryChooser v) {
+				if (!StrUtil.isEmpty(val)) {
+					v.setInitialDirectory(new File(val));
+				}
+			} else if (value instanceof TextInputControl v) {
+				v.setText(val);
+			} else if (value instanceof CheckBox v) {
+				v.setSelected(Boolean.parseBoolean(val));
+			} else if (value instanceof DatePicker v) {
+				if (!StrUtil.isEmpty(val)) {
+					v.setValue(LocalDate.parse(val));
+				}
+			} else {
+				// do nothing
+			}
+		});
+	}
 
-        final VBox rightPanel = new VBox();
-        rightPanel.getStyleClass().add("right-panel");
-        rightPanel.setMaxHeight(Double.MAX_VALUE);
+	/**
+	 * Manually call the current method after the function completes For example
+	 * manually calling the current method after clicking the generate button
+	 */
+	protected void bindUserData() {
+		if (userData.size() == 0 || StrUtil.isEmpty(getSampleId())) {
+			return;
+		}
+		FxXmlUtil.set(getSampleXmlPrefix(), "id", getSampleId());
+		FxXmlUtil.set(getSampleXmlPrefix(), "version", getSampleVersion());
+		userData.forEach((key, value) -> {
+			String k = getSampleXmlPrefix() + "." + key;
+			if (value instanceof FileChooser v) {
+				FxXmlUtil.set(k, v.getInitialDirectory());
+			} else if (value instanceof DirectoryChooser v) {
+				FxXmlUtil.set(k, v.getInitialDirectory());
+			} else if (value instanceof TextInputControl v) {
+				FxXmlUtil.set(k, v.getText());
+			} else if (value instanceof CheckBox v) {
+				FxXmlUtil.set(k, v.isSelected());
+			} else if (value instanceof DatePicker v) {
+				FxXmlUtil.set(k, v.getValue());
+			} else {
+				// do nothing
+			}
+		});
+	}
 
-        boolean addRightPanel = false;
+	protected String getSampleXmlPrefix() {
+		return getProjectName() + "." + getSampleId();
+	}
 
-        Label sampleName = new Label(sample.getSampleName());
-        sampleName.getStyleClass().add("sample-name");
-        rightPanel.getChildren().add(sampleName);
+	/**
+	 * Utility method to create the default look for samples.
+	 */
+	public static Node buildSample(Sample sample, Stage stage) {
+		SplitPane splitPane = new SplitPane();
+		// we guarantee that the build order is panel then control panel.
+		final Node samplePanel = sample.getPanel(stage);
+		final Node controlPanel = sample.getControlPanel();
+		sample.initialize();
+		splitPane.setDividerPosition(0, sample.getControlPanelDividerPosition());
 
-        // --- project name & version
-        String version = sample.getProjectVersion();
-        version = version == null ? "" : "@version@".equals(version) ? "" : " " + version.trim();
+		if (samplePanel != null) {
+			samplePanel.getStyleClass().add("center-panel");
+			splitPane.getItems().add(samplePanel);
+		}
 
-        final String projectName = sample.getProjectName() + version;
-        if (!projectName.isEmpty()) {
-            Label projectNameTitleLabel = new Label(I18nUtils.get("frame.sample.rightPanel.project") + ": ");
-            projectNameTitleLabel.getStyleClass().add("project-name-title");
+		final VBox rightPanel = new VBox();
+		rightPanel.getStyleClass().add("right-panel");
+		rightPanel.setMaxHeight(Double.MAX_VALUE);
 
-            Label projectNameLabel = new Label(projectName);
-            projectNameLabel.getStyleClass().add("project-name");
-            projectNameLabel.setWrapText(true);
+		boolean addRightPanel = false;
 
-            TextFlow textFlow = new TextFlow(projectNameTitleLabel, projectNameLabel);
-            rightPanel.getChildren().add(textFlow);
-        }
+		Label sampleName = new Label(sample.getSampleName());
+		sampleName.getStyleClass().add("sample-name");
+		rightPanel.getChildren().add(sampleName);
 
-        // --- sample version
-        final String sampleVersion = sample.getSampleVersion();
-        if (sampleVersion != null && !sampleVersion.isEmpty()) {
-            Label sampleVersionLabel = new Label(
-                    I18nUtils.get("frame.sample.rightPanel.sampleVersion") + ": " + sampleVersion);
-            sampleVersionLabel.getStyleClass().add("sampleVersion");
-            sampleVersionLabel.setWrapText(true);
-            rightPanel.getChildren().add(sampleVersionLabel);
-            addRightPanel = true;
-        }
+		// --- project name & version
+		String version = sample.getProjectVersion();
+		version = version == null ? "" : "@version@".equals(version) ? "" : " " + version.trim();
 
-        // --- description
-        final String description = sample.getSampleDescription();
-        if (description != null && !description.isEmpty()) {
-            Label descriptionLabel = new Label(description);
-            descriptionLabel.getStyleClass().add("description");
-            descriptionLabel.setWrapText(true);
-            rightPanel.getChildren().add(descriptionLabel);
+		final String projectName = sample.getProjectName() + version;
+		if (!projectName.isEmpty()) {
+			Label projectNameTitleLabel = new Label(I18nUtils.get("frame.sample.rightPanel.project") + ": ");
+			projectNameTitleLabel.getStyleClass().add("project-name-title");
 
-            addRightPanel = true;
-        }
+			Label projectNameLabel = new Label(projectName);
+			projectNameLabel.getStyleClass().add("project-name");
+			projectNameLabel.setWrapText(true);
 
-        if (controlPanel != null) {
-            rightPanel.getChildren().add(new Separator());
+			TextFlow textFlow = new TextFlow(projectNameTitleLabel, projectNameLabel);
+			rightPanel.getChildren().add(textFlow);
+		}
 
-            controlPanel.getStyleClass().add("control-panel");
-            rightPanel.getChildren().add(controlPanel);
-            VBox.setVgrow(controlPanel, Priority.ALWAYS);
-            addRightPanel = true;
-        }
+		// --- sample version
+		final String sampleVersion = sample.getSampleVersion();
+		if (sampleVersion != null && !sampleVersion.isEmpty()) {
+			Label sampleVersionLabel = new Label(
+					I18nUtils.get("frame.sample.rightPanel.sampleVersion") + ": " + sampleVersion);
+			sampleVersionLabel.getStyleClass().add("sampleVersion");
+			sampleVersionLabel.setWrapText(true);
+			rightPanel.getChildren().add(sampleVersionLabel);
+			addRightPanel = true;
+		}
 
-        if (addRightPanel) {
-            ScrollPane scrollPane = new ScrollPane(rightPanel);
-            scrollPane.setMaxHeight(Double.MAX_VALUE);
-            scrollPane.setFitToWidth(true);
-            scrollPane.setFitToHeight(true);
-            SplitPane.setResizableWithParent(scrollPane, false);
-            splitPane.getItems().add(scrollPane);
-        }
+		// --- description
+		final String description = sample.getSampleDescription();
+		if (description != null && !description.isEmpty()) {
+			Label descriptionLabel = new Label(description);
+			descriptionLabel.getStyleClass().add("description");
+			descriptionLabel.setWrapText(true);
+			rightPanel.getChildren().add(descriptionLabel);
 
-        return splitPane;
-    }
+			addRightPanel = true;
+		}
 
-    @Override
-    public void dispose() {
-    }
+		if (controlPanel != null) {
+			rightPanel.getChildren().add(new Separator());
+
+			controlPanel.getStyleClass().add("control-panel");
+			rightPanel.getChildren().add(controlPanel);
+			VBox.setVgrow(controlPanel, Priority.ALWAYS);
+			addRightPanel = true;
+		}
+
+		if (addRightPanel) {
+			ScrollPane scrollPane = new ScrollPane(rightPanel);
+			scrollPane.setMaxHeight(Double.MAX_VALUE);
+			scrollPane.setFitToWidth(true);
+			scrollPane.setFitToHeight(true);
+			SplitPane.setResizableWithParent(scrollPane, false);
+			splitPane.getItems().add(scrollPane);
+		}
+
+		return splitPane;
+	}
+
+	@Override
+	public void dispose() {
+	}
 }
