+++ smc/src/main/java/com/tlcsdm/smc/SmcSample.java
@@ -1,24 +1,9 @@
 package com.tlcsdm.smc;
 
-import java.io.File;
-import java.time.LocalDate;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
 import com.tlcsdm.core.javafx.util.Config;
-import com.tlcsdm.core.javafx.util.FxXmlUtil;
 import com.tlcsdm.frame.SampleBase;
 
-import cn.hutool.core.util.StrUtil;
-import javafx.scene.control.CheckBox;
-import javafx.scene.control.DatePicker;
-import javafx.scene.control.TextInputControl;
-import javafx.stage.DirectoryChooser;
-import javafx.stage.FileChooser;
-
 public abstract class SmcSample extends SampleBase {
-
-	protected Map<String, Object> userData = new LinkedHashMap<>();
 
 	@Override
 	public String getProjectName() {
@@ -28,76 +13,6 @@ public abstract class SmcSample extends SampleBase {
 	@Override
 	public String getProjectVersion() {
 		return Config.JAVAFX_TOOL_VERSION;
-	}
-
-	@Override
-	public void initialize() {
-		initializeUserData();
-	}
-
-	/**
-	 * Because initialize() is called after getPanel() so userData needs to be
-	 * initialized in getPanel()
-	 */
-	protected void initializeUserData() {
-		if (!FxXmlUtil.hasKey(getSampleXmlPrefix(), "id")) {
-			return;
-		}
-		userData.forEach((key, value) -> {
-			String k = getSampleXmlPrefix() + "." + key;
-			String val = FxXmlUtil.get(k, "");
-			if (value instanceof FileChooser v) {
-				if (!StrUtil.isEmpty(val)) {
-					v.setInitialDirectory(new File(val));
-				}
-			} else if (value instanceof DirectoryChooser v) {
-				if (!StrUtil.isEmpty(val)) {
-					v.setInitialDirectory(new File(val));
-				}
-			} else if (value instanceof TextInputControl v) {
-				v.setText(val);
-			} else if (value instanceof CheckBox v) {
-				v.setSelected(Boolean.parseBoolean(val));
-			} else if (value instanceof DatePicker v) {
-				if (!StrUtil.isEmpty(val)) {
-					v.setValue(LocalDate.parse(val));
-				}
-			} else {
-				// do nothing
-			}
-		});
-	}
-
-	/**
-	 * Manually call the current method after the function completes For example
-	 * manually calling the current method after clicking the generate button
-	 */
-	protected void bindUserData() {
-		if (userData.size() == 0 || StrUtil.isEmpty(getSampleId())) {
-			return;
-		}
-		FxXmlUtil.set(getSampleXmlPrefix(), "id", getSampleId());
-		FxXmlUtil.set(getSampleXmlPrefix(), "version", getSampleVersion());
-		userData.forEach((key, value) -> {
-			String k = getSampleXmlPrefix() + "." + key;
-			if (value instanceof FileChooser v) {
-				FxXmlUtil.set(k, v.getInitialDirectory());
-			} else if (value instanceof DirectoryChooser v) {
-				FxXmlUtil.set(k, v.getInitialDirectory());
-			} else if (value instanceof TextInputControl v) {
-				FxXmlUtil.set(k, v.getText());
-			} else if (value instanceof CheckBox v) {
-				FxXmlUtil.set(k, v.isSelected());
-			} else if (value instanceof DatePicker v) {
-				FxXmlUtil.set(k, v.getValue());
-			} else {
-				// do nothing
-			}
-		});
-	}
-
-	protected String getSampleXmlPrefix() {
-		return getProjectName() + "." + getSampleId();
 	}
 
 }
